requirements.txt
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
Pillow==10.1.0
matplotlib==3.8.2
pandas==2.1.4
openpyxl==3.1.2
PyPDF2==3.0.1
